samples/python/requirements.txt
Flask==0.5
requests==2.18.0
PyYAML==5.1
